requirements.txt
fastapi
uvicorn
wheel
sqlalchemy
zope.sqlalchemy
psycopg2-binary 
psycopg2
passlib 
bcrypt==4.0.1
pydantic 
python-jose
databases
asyncpg
jinja2
python-multipart
aiosqlite
